.unpublished/python-ai-helper/files/pyproject.toml
[project]
name = "ai-maintenance-agent"
version = "0.1.0"
description = "Elite AI-powered Python maintenance agent"
requires-python = ">=3.13"
dependencies = [
    "libcst>=1.5.0",
    "requests>=2.32.0",
    "pyyaml>=6.0.1",
]

[project.optional-dependencies]
dev = [
    "ruff>=0.8.0",
    "mypy>=1.13.0",
    "pytest>=8.3.0",
]

[tool.ruff]
line-length = 100
target-version = "py313"
# Redirect Ruff cache to ephemeral tmpfs
cache-dir = "/tmp/.ruff_cache"

[tool.ruff.lint]
select = ["E", "F", "I", "N", "UP", "B", "C4"]

[tool.mypy]
python_version = "3.13"
strict = true
ignore_missing_imports = true

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = "test_*.py"
